feat(core): restore platform-agnostic sidecar and ingest HMAC
Reintegrate the VPS-proven tree (no in-repo connectors) onto 1.3.1, keeping version, stats migration, and instance-info.
