+++ .env.example
@@ -1,11 +1,4 @@
 # ---------------------------------------------------------
-# Streaming Platform Connector Configuration
-# ---------------------------------------------------------
-# Name of the active connector module (e.g. jellyfin, peertube, owncast, etc.)
-ACTIVE_CONNECTOR=your_connector_name_here
-UPSTREAM_URL=http://localhost:your_port_here
-
-# ---------------------------------------------------------
 # Circle User-Controlled Wallets (Required)
 # If you don't know how to get these, follow the guides:
 # - API Key: docs/tutorials/circle-api-key.md
@@ -18,45 +11,27 @@ CIRCLE_APP_ID=your_app_id_here
 CIRCLE_GOOGLE_CLIENT_ID=your_google_web_client_id
 # Optional:
 # CIRCLE_FACEBOOK_APP_ID=your_facebook_app_id
-# Google OAuth authorized redirect URI must include your PeerTube origin (e.g. https://peertube.example.com).
+# Google OAuth authorized redirect URI must include the origin where viewers open the player.
 
 # ---------------------------------------------------------
-# Security Configuration
+# Security
 # ---------------------------------------------------------
-# High-entropy secret key used to encrypt ephemeral session keys at rest.
-# Must be a secure random string (e.g. 32+ characters).
+# Encrypts ephemeral session keys at rest (32+ random characters).
 MASTER_KEY=your_secure_master_key_here
 
-# ---------------------------------------------------------
-# Arc Testnet RPC Node (Optional — Defaults to https://rpc.testnet.arc.network)
-# ---------------------------------------------------------
-# The RPC endpoint for Arc Testnet. You can use the official public endpoint,
-# or supply a custom endpoint URL from Canteen, QuickNode, dRPC, or Alchemy.
-# ARC_RPC_URL=https://rpc.testnet.arc.network
-# ARC_RPC_URL=https://rpc.testnet.arc-node.thecanteenapp.com/v1/your-api-key
-# ARC_RPC_URL=https://your-node.arc-testnet.quiknode.pro/your-api-key/
-ARC_RPC_URL=https://rpc.testnet.arc.network
+# HMAC for POST /api/core/v1/sessions/start and /stop (required).
+# Plugins send X-Tessera-Timestamp, X-Tessera-Nonce, X-Tessera-Signature.
+# Tessera will not start if this is missing.
+TESSERA_INGEST_SECRET=your_ingest_hmac_secret_here
 
+# Optional Arc Testnet RPC. If unset, defaults to https://rpc.testnet.arc.network
+# ARC_RPC_URL=
 
-# ---------------------------------------------------------
-# Connector Webhook Security
-# ---------------------------------------------------------
-# HMAC secret for each active connector's webhook endpoint.
-# Pattern: TESSERA_CONNECTOR_SECRET_<NAME_UPPERCASE>
-# PeerTube plugin setting "webhook-secret" must equal TESSERA_CONNECTOR_SECRET_PEERTUBE.
-# Legacy PEERTUBE_WEBHOOK_SECRET is bridged at PeerTube connector startup if the new var is unset.
-# Example:
-# TESSERA_CONNECTOR_SECRET_PEERTUBE=your_secret_here
-# TESSERA_CONNECTOR_SECRET_JELLYFIN=your_secret_here
-
-# Public URL for the sidecar API (Defaults to http://localhost:7878)
+# Public URL for the sidecar API (default http://localhost:7878)
 PUBLIC_URL=http://localhost:7878
 
-# Server port (Defaults to 7878)
+# Server port (default 7878)
 PORT=7878
 
-# Minimum balance the ephemeral wallet must hold to initiate a session
 # MIN_WALLET_BALANCE=0.01
-
-# The amount of USDC reserved for gas during withdrawal settlement on Arc Testnet (Defaults to 0.01)
 # RETAINED_GAS_AMOUNT=0.01
