+++ package.json
@@ -4,7 +4,6 @@
   "description": "Per-second streaming payments for self-hosted platforms, powered by Circle x402.",
   "main": "src/index.ts",
   "scripts": {
-    "setup": "node scripts/setup.js",
     "start": "node dist/index.js",
     "dev": "ts-node src/index.ts",
     "build": "tsc",
@@ -20,8 +19,7 @@
     "usdc",
     "streaming",
     "micropayments",
-    "sidecar",
-    "owncast"
+    "sidecar"
   ],
   "author": "JaDi03",
   "license": "Apache-2.0",
